fix: ci issues
