+++ apps/browser-extension-wallet/src/hooks/index.ts
@@ -19,4 +19,5 @@ export * from './useStakingRewards';
 export * from './useMaxAda';
 export * from './useSyncingTheFirstTime';
 export * from './useGetHandles';
+export * from './useHandleResolver';
 export * from './useAdaHandle';
